Add message dialog to confirm deletion.
Add instance of IMessageDialogService to FriendDetailViewModel and ask user if they are sure about deletion before executing logic in OnDeleteExecute.

+++ FriendOrganizer.UI/ViewModel/FriendDetailViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using FriendOrganizer.Model;
 using FriendOrganizer.UI.Data.Repositories;
 using FriendOrganizer.UI.Event;
+using FriendOrganizer.UI.View.Services;
 using FriendOrganizer.UI.Wrapper;
 using Prism.Commands;
 using Prism.Events;
@@ -11,17 +12,20 @@ namespace FriendOrganizer.UI.ViewModel
 {
     public class FriendDetailViewModel : ViewModelBase, IFriendDetailViewModel
     {
-        private IFriendRepository _friendRepository;
-        private IEventAggregator _eventAggregator;
-        private FriendWrapper _friend;
         private bool _hasChanges;
+        private FriendWrapper _friend;
+        private IEventAggregator _eventAggregator;
+        private IFriendRepository _friendRepository;
+        private IMessageDialogService _messageDialogService;
 
         public FriendDetailViewModel(
             IFriendRepository friendRepository,
-            IEventAggregator eventAggregator)
+            IEventAggregator eventAggregator,
+            IMessageDialogService messageDialogService)
         {
             _friendRepository = friendRepository;
             _eventAggregator = eventAggregator;
+            _messageDialogService = messageDialogService;
 
             SaveCommand = new DelegateCommand(OnSaveExecute, OnSaveCanExecute);
             DeleteCommand = new DelegateCommand(OnDeleteExecute);
@@ -101,10 +105,17 @@ namespace FriendOrganizer.UI.ViewModel
 
         private async void OnDeleteExecute()
         {
-            _friendRepository.Remove(Friend.Model);
-            await _friendRepository.SaveAsync();
-            _eventAggregator.GetEvent<AfterFriendDeletedEvent>()
-                .Publish(Friend.Id);
+            var result = _messageDialogService.ShowOKCancelDialog(
+                $"Do you really want to delete the friend {Friend.FirstName} {Friend.LastName}?",
+                "Question");
+
+            if (result == MessageDialogResult.OK)
+            {
+                _friendRepository.Remove(Friend.Model);
+                await _friendRepository.SaveAsync();
+                _eventAggregator.GetEvent<AfterFriendDeletedEvent>()
+                    .Publish(Friend.Id);
+            }
         }
 
         private Friend CreateNewFriend()
